Revert [r3709] in favour of a fix that preserves the use `off_t` for file offsets.
Ticket:		#568

git-svn-id: svn://svn.code.sf.net/p/elftoolchain/code/trunk@3710 95820547-d848-0410-985e-9ae8fe0fa350

+++ libelf/elf_next.c
@@ -35,7 +35,7 @@ ELFTC_VCSID("$Id$");
 Elf_Cmd
 elf_next(Elf *e)
 {
-	size_t next;
+	off_t next;
 	Elf *parent;
 
 	if (e == NULL)
@@ -50,14 +50,15 @@ elf_next(Elf *e)
 	assert(parent->e_cmd == ELF_C_READ);
 	assert(e->e_rawfile > parent->e_rawfile);
 
-	next = e->e_rawfile - parent->e_rawfile + e->e_rawsize;
+	next = e->e_rawfile - parent->e_rawfile + (off_t) e->e_rawsize;
 	next = (next + 1) & ~1;	/* round up to an even boundary */
 
 	/*
 	 * Setup the 'e_next' field of the archive descriptor for the
 	 * next call to 'elf_begin()'.
 	 */
-	parent->e_u.e_ar.e_next = (next >= parent->e_rawsize) ?  0 : next;
+	parent->e_u.e_ar.e_next = (next >= (off_t) parent->e_rawsize) ?
+	    (off_t) 0 : next;
 
 	/*
 	 * Return an error if the 'e_next' field falls outside the current
@@ -68,7 +69,7 @@ elf_next(Elf *e)
 	 * traversal of a too-small archive even if client code forgets to
 	 * check the return value from elf_next(3).
 	 */
-	if (next > parent->e_rawsize) {
+	if (next > (off_t) parent->e_rawsize) {
 		LIBELF_SET_ERROR(ARGUMENT, 0);
 		return (ELF_C_NULL);
 	}
